+++ lib/utils/ticket-form-bootstrap-server-cache.ts
@@ -13,6 +13,7 @@ const EDIT_LOOKUP_CATEGORIES = [
   "sides",
   "roll_direction",
   "finishing",
+  "designer",
   "ticket_priority",
   "quote_channel",
   "ticket_payment",